+++ cores/arduino/stm32/stm32_eeprom.h
@@ -55,13 +55,13 @@
 /* Exported macro ------------------------------------------------------------*/
 /* Exported functions ------------------------------------------------------- */
 
-uint8_t eeprom_read_byte(const uint16_t pos);
-void eeprom_write_byte(uint16_t pos, uint8_t value);
+uint8_t eeprom_read_byte(const uint32_t pos);
+void eeprom_write_byte(uint32_t pos, uint8_t value);
 
 void eeprom_buffer_fill();
 void eeprom_buffer_flush();
-uint8_t eeprom_buffered_read_byte(const uint16_t pos);
-void eeprom_buffered_write_byte(uint16_t pos, uint8_t value);
+uint8_t eeprom_buffered_read_byte(const uint32_t pos);
+void eeprom_buffered_write_byte(uint32_t pos, uint8_t value);
 
 #ifdef __cplusplus
 }
